+++ eth_defi/erc_4626/scan.py
@@ -17,7 +17,7 @@ def create_vault_scan_record(
     web3: Web3,
     detection: ERC4262VaultDetection,
     block_identifier: BlockIdentifier,
-):
+) -> dict:
     """Create a row in the result table.
 
     - Connect to the chain to read further vault metadata via JSON-RPC calls
@@ -37,13 +37,28 @@ def create_vault_scan_record(
             "Protocol": "<unknown>",
             "Denomination": "",
             "NAV": 0,
+            "Mgmt fee": "-",
+            "Perf fee": "-",
             "Shares": 0,
             "First seen": detection.first_seen_at,
+            "_detection_data": detection,
         }
     else:
         # Try to figure out the correct vault subclass
         # to pull out the data like fees
         vault = cast(ERC4626Vault, vault)
+        try:
+            management_fee = vault.get_management_fee(block_identifier)
+            assert type(management_fee) == float, f"Vault {vault} gave {management_fee}"
+        except NotImplementedError:
+            management_fee = None
+
+        try:
+            performance_fee = vault.get_performance_fee(block_identifier)
+            assert type(performance_fee) == float
+        except NotImplementedError:
+            performance_fee = None
+
         data = {
             "Symbol": vault.symbol,
             "Name": vault.name,
@@ -51,8 +66,11 @@ def create_vault_scan_record(
             "Denomination": vault.denomination_token.symbol,
             "NAV": vault.fetch_total_assets(block_identifier),
             "Protocol": get_vault_protocol_name(detection.features),
+            "Mgmt fee": management_fee,
+            "Perf fee": performance_fee,
             "Shares": vault.fetch_total_supply(block_identifier),
             "First seen": detection.first_seen_at,
+            "_detection_data": detection,
         }
 
     return data
@@ -66,7 +84,7 @@ def create_vault_scan_record_subprocess(
     web3factory: Web3Factory,
     detection: ERC4262VaultDetection,
     block_number: int,
-):
+) -> dict:
     """Process remaining vault data reads using multiprocessing
 
     - See :py:func:`create_vault_scan_record`