+++ eth_defi/morpho/__init__.py
@@ -1,0 +1,5 @@
+"""Morpho vault protocol support.
+
+- `See an example vault here <https://app.gauntlet.xyz/vaults/eth:0x4881ef0bf6d2365d3dd6499ccd7532bcdbce0658>`__
+- `Example contract <https://basescan.org/address/0x6b13c060F13Af1fdB319F52315BbbF3fb1D88844#readContract>`__
+"""