+++ scripts/erc-4626/scan-vaults.py
@@ -3,6 +3,7 @@
 - Set up a HyperSync based vault discovery client
 - As the writing of this, we get 1108 leads on Base
 - Takes environment variables ``JSON_RPC_URL``, ``LOG_LEVEL``, ``END_BLOCK``
+- Save data to /tmp: both raw Python objects and Parquet dump
 
 Usage:
 
@@ -19,6 +20,7 @@ Or for faster small sample scan limit the end block:
 import decimal
 import logging
 import os
+import pickle
 import sys
 from decimal import Decimal
 from pathlib import Path
@@ -30,6 +32,7 @@ from joblib import Parallel, delayed
 
 from tqdm_loggable.auto import tqdm
 
+from eth_defi.erc_4626.core import ERC4626Feature
 from eth_defi.erc_4626.hypersync_discovery import HypersyncVaultDiscover
 from eth_defi.erc_4626.scan import create_vault_scan_record_subprocess
 from eth_defi.hypersync.server import get_hypersync_server
@@ -78,6 +81,10 @@ def main():
     else:
         end_block = int(end_block)
 
+    output_folder = os.environ.get("OUTPUT_FOLDER")
+    if output_folder is None:
+        output_folder = "/tmp"
+
     # Create a scanner that uses web3, HyperSync and subprocesses
     vault_discover = HypersyncVaultDiscover(
         web3,
@@ -95,41 +102,83 @@ def main():
     logger.info("Extracting remaining vault metadata for %d vaults", len(vault_detections))
 
     # Quite a mouthful line to create a row of output for each vault detection using subproces pool
-    desc = f"Extracting vault metadata"
+    desc = f"Extracting vault metadata using {max_workers} workers"
     rows = worker_processor(delayed(create_vault_scan_record_subprocess)(web3factory, d, end_block) for d in tqdm(vault_detections, desc=desc))
 
     print(f"Total {len(rows)} vaults detected")
     df = pd.DataFrame(rows)
-
+    # Cannot export the raw Python object,
+    # this is for the pickle only
+    df = df.drop(columns="_detection_data")
     df = df.sort_values("First seen")
 
-    output_fname = Path("/tmp/vaults.parquet")
+    #
+    # Save raw data rows
+    #
+
+    chain = web3.eth.chain_id
+    output_fname = Path(f"{output_folder}/chain-{chain}-vaults.parquet")
+    parquet_df = df.copy()
+    parquet_df = parquet_df.fillna(pd.NA)  # fillna replaces None and NaN with pd.NA
+    parquet_df['Mgmt fee'] = pd.to_numeric(parquet_df['Mgmt fee'], errors='coerce')
+    parquet_df['Perf fee'] = pd.to_numeric(parquet_df['Perf fee'], errors='coerce')
     print(f"Saving raw data to {output_fname}")
-    df.to_parquet(output_fname)
+    parquet_df.to_parquet(output_fname)
+
+    #
+    # Save machine-readable output
+    #
+
+    # Save dict -> data mapping with raw data to be read in notebooks and such.
+    # This will preserve raw vault detection objects.
+    data_dict = {r["Address"]: r for r in rows}
+    output_fname = Path(f"{output_folder}/chain-{chain}-vaults.pickle")
+    print(f"Saving raw data to {output_fname}")
+    pickle.dump(data_dict, output_fname.open("wb"))
+
+    #
+    # Display in terminal
+    #
 
     # Format DataFrame output for terminal
     df["First seen"] = df["First seen"].dt.strftime('%Y-%b-%d')
+    df["Mgmt fee"] = df["Mgmt fee"].apply(lambda x: f"{x:.1%}" if type(x) == float else "-")
+    df["Perf fee"] = df["Perf fee"].apply(lambda x: f"{x:.1%}" if type(x) == float else "-")
     # df["Address"] = df["Address"].apply(lambda x: x[0:8])  # Address is too wide in terminal
     df = df.set_index("Address")
 
     # Round dust to zero, drop to 4 decimals
     def round_below_epsilon(x, epsilon=Decimal("0.1"), round_factor=Decimal("0.001")):
         if isinstance(x, Decimal):
+
+            # Eliminate dust
             x = Decimal('0') if abs(x) < epsilon else x
-            if x < 1000:
+
+            float_x = float(x)
+
+            # Get rid of numbers with too many digits
+            if float_x >= 1e12:  # Trillions
+                return f"{float_x / 1e12:.1f}T"
+            elif float_x >= 1e9:  # Billions
+                return f"{float_x / 1e9:.1f}G"
+            elif float_x >= 1e6:  # Millions
+                return f"{float_x / 1e6:.1f}M"
+            elif float_x >= 1e3:  # Millions
+                return f"{float_x / 1e6:.1f}K"
+            else:
                 try:
                     x = x.quantize(round_factor)
                 except decimal.InvalidOperation:
                     logger.warning("Cannot quantise: %s", x)
-            else:
-                # Decimals are not important in large values
-                x = int(x)
-            return x
 
         return x  # Not decimal
 
     # Apply the function to all elements in the DataFrame
     df = df.apply(lambda col: col.map(round_below_epsilon))
+
+    erc_7540s = [v for v in rows if ERC4626Feature.erc_7540_like in v["_detection_data"].features]
+    print(f"Total: {len(df)} vaults detected")
+    print(f"ERC-7540: {len(erc_7540s)} vaults detected")
 
     with pd.option_context('display.max_rows', None):
         display(df)